bridgeql: fix fail if exception is thrown
Incase when an exception is thrown the application throws another error since the exception object was directly assigned to the message field of response  which the JSONEncoder was unable to encode into json. This was fixed by assigning error message to the message field of response in case of any exception.

Additional changes: added status code of 400 in case of any error.