skip property name validation
Currently we validate property name for a model by checking if the model has
the attribute with name that of property. This is causing some issues as
there might be a filter such as fkey_isnull, now since isnull is not a valid
property name this will cause an InvalidModelFieldName error.
Hence we decided to skip this validation and only check for restrictions and let
django handle invalid property/fieldname

Added test cases

Signed-off-by: Priyank Singh <priyanksi@vmware.com>