Use select_related if query contains foreign key references
Currently we are directly making a db query if we get a property of a model which is
declared as a foreign key. Example `os.name`. This makes the data retrieval very slow
if we have many foreign key references since we execute seperate db queries.
To avoid that we use select_related on queryset by default if we have foreign_key
in the query.

Additional changes: improve logging
Fixes: #13
Signed-off-by: Priyank Singh <priyanksi@vmware.com>