perform limit and offset operation before evaluating queryset
Incase of foreign key reference, we evaluate the complete queryset by iterating through
all of the rows first and then perform slicing operation for limit and offset.
Hence in cases if limit is present and total number of rows in a table is very large
`__add_fields` will iterate through all the rows, which is very inefficient.
So this change limits the query first before evaluating the query set.
This will make queries containing foreign key faster if limit is present

Additional changes: remove unused imports

Signed-off-by: Priyank Singh <priyanksi@vmware.com>