Implemented settings BRIDGEQL_ALLOWED_APPS (#21)
* Implemented settings BRIDGEQL_ALLOWED_APPS

To see the schema of project, user need to define BRIDGEQL_ALLOWED_APPS
in settings.py. If this settings is not defined then all the local apps
(app created by user) will be in the allowed apps list. User must define
either BASE_DIR or SITE_ROOT or BRIDGEQL_ALLOWED_APPS. In nothing is
defined, then a message will be displayed in schema page to initilize
BRIDGEQL_ALLOWED_APPS

* Typo fixes and using bridgeql settings for allowed_apps

By default value of BRIDGEQL_ALLOWED_APPS is set to a empty list. The
value of BRIDGEQL_ALLOWED_APPS can be defined in project's settings.py
If not defined, it will be by default empty list.

Signed-off-by: Vidya Sagar <vidya.sagar0911@gmail.com>