writer: add writer API to create or update django models
This change includes a new urls which writes (create or update) the django model
based on the parameters passed such as db_name or pk. If a primary key is included in
the url then bridgeql will try to update the model, else it will create it.

Added test cases.

Signed-off-by: Priyank Singh <priyanksi@vmware.com>