Writer api (#18)
* writer: add writer API to create or update django models

This change includes a new urls which writes (create or update) the django model
based on the parameters passed such as db_name or pk. If a primary key is included in
the url then bridgeql will try to update the model, else it will create it.

While updating a model if primary key used to get the model is invalid,
then bridgeql raises an InvalidPKException.

Inorder to support patch method, we need to retrieve payload
directly from request body since there is no equivalent
`request.POST` for patch method. However this introduces a constraint
that the client should always pass the request body in json format
with content type `applicaton/json`.
Added a better way of handling exception

use separate auth for reader and writer, `BRIDGEQL_AUTHENTICATION_DECORATOR` now accepts,
a dictionary with attributes reader and writer containing their respective auth decorator import path.

URLs are now of the pattern create/, read/, update/. New url of delete would also be added
which would delete the resource based on the primary key.


* update version to 0.2.0
---------

Signed-off-by: Piyus Kumar <piyusk@vmware.com>
Signed-off-by: Priyank Singh <priyanksi@vmware.com>
Co-authored-by: Priyank Singh <priyanksi@vmware.com>