add stream API
Add stream API to stream data for a large number of results,
this can be controlled by chunk_size query parameter,
the server will sleep for 1 second when it has yieled `chunk_size` data.
The server needs to sleep for 1 second since it needs time to gather the records (chunk_size).

Signed-off-by: Priyank Singh <priyanksi@vmware.com>